% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R, R/myfridgebuddydata-package.R
\docType{package}
\name{myfridgebuddydata}
\alias{myfridgebuddydata}
\title{myfridgebuddydata: EpiRecipes Data from Kaggle}
\source{
EpiRecipes Kaggle dataset: \url{https://www.kaggle.com/datasets/hugodarwood/epirecipes}

EpiRecipes Kaggle dataset: \url{https://www.kaggle.com/datasets/hugodarwood/epirecipes}
}
\description{
A data-only package containing the EpiRecipes dataset sourced from Kaggle.
It includes recipe metadata, categories, ingredients, and directions.

A data-only package containing the EpiRecipes dataset sourced from Kaggle. It includes recipe metadata,
categorical associations, ingredient listings, and preparation directions for 20k+ recipes.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
